Separate core drawinglayer func. into drawinglayercore library
This separates the drawinglayer core functionallity into a
separate library, to keep a strict separation what is backend
dependent and what is not. More strict separation can be done
at a later date.

This will make it possible to push part of drawinglayer
(part of processor2d) directly into VCL.

Change-Id: Ibc26580067e50bf20d7cdd37fa0e44eb10200878