Fix sc tiledrendering unit test document-setup
The calc tiledrendering unit tests that use view-data's nPPTX directy or
indirectly are failing in pristine master build (at least in my setup).

When creating a document for each test via createDoc(), the docshell has
an output-factor != 1.0 by default (at least for these unit-tests). We
do call ScModelObj::initializeForTiledRendering() soon afterwards which
does the right thing by setting it back to 1.0 (indirectly), but nPPTX
value in view-data is not updated and stays wrong.

I haven't had a chance to investigate more why this happens only in my
setup, but as an interim fix, an explicit call to RefreshZoom() is added
to createDoc().

Change-Id: Ia7044381fa203e03c56f8a2589140c634347fe66